feat(avo-1910): pupil search page tab
https://meemoo.atlassian.net/browse/AVO-1910

still todo: handle link to series and links to fragments/collections inside the same page
