+++ src/admin/collectionsOrBundles/collections-or-bundles.const.ts
@@ -1,7 +1,6 @@
-import { DocumentNode } from 'graphql';
-
 import { ButtonType, SelectOption } from '@viaa/avo2-components';
 import { Avo } from '@viaa/avo2-types';
+import type { DocumentNode } from 'graphql';
 
 import { CheckboxDropdownModalProps, CheckboxOption } from '../../shared/components';
 import { BooleanCheckboxDropdownProps } from '../../shared/components/BooleanCheckboxDropdown/BooleanCheckboxDropdown';
@@ -32,11 +31,9 @@ export const COLLECTIONS_OR_BUNDLES_PATH = {
 
 export const ITEMS_PER_PAGE = 10;
 
-export const TABLE_COLUMN_TO_DATABASE_ORDER_OBJECT: Partial<
-	{
-		[columnId in CollectionTableCols]: (order: Avo.Search.OrderDirection) => any;
-	}
-> = {
+export const TABLE_COLUMN_TO_DATABASE_ORDER_OBJECT: Partial<{
+	[columnId in CollectionTableCols]: (order: Avo.Search.OrderDirection) => any;
+}> = {
 	owner_profile_id: (order: Avo.Search.OrderDirection) => ({
 		profile: { usersByuserId: { last_name: order } },
 	}),
@@ -80,11 +77,9 @@ export const TABLE_COLUMN_TO_DATABASE_ORDER_OBJECT: Partial<
 	}),
 };
 
-export const EDITORIAL_TABLE_COLUMN_TO_DATABASE_ORDER_OBJECT: Partial<
-	{
-		[columnId in CollectionTableCols]: (order: Avo.Search.OrderDirection) => any;
-	}
-> = {
+export const EDITORIAL_TABLE_COLUMN_TO_DATABASE_ORDER_OBJECT: Partial<{
+	[columnId in CollectionTableCols]: (order: Avo.Search.OrderDirection) => any;
+}> = {
 	...TABLE_COLUMN_TO_DATABASE_ORDER_OBJECT,
 	owner_profile_id: (order: Avo.Search.OrderDirection) => ({
 		owner: { profile: { usersByuserId: { last_name: order } } },
