+++ src/assignment/assignment.types.ts
@@ -2,6 +2,8 @@ import { TableColumn } from '@viaa/avo2-components';
 import { Avo } from '@viaa/avo2-types';
 import { AssignmentBlock, AssignmentLabel_v2 } from '@viaa/avo2-types/types/assignment';
 
+import { FilterState } from '../search/search.types';
+
 export type AssignmentOverviewTableColumns =
 	| 'title'
 	| 'assignment_type'
@@ -80,3 +82,7 @@ export type AssignmentFormState = Pick<Avo.Assignment.Assignment_v2, 'title'> &
 	};
 
 export type AssignmentBlockTypeDict<T> = { [key in AssignmentBlockType]: T }; // eslint-disable-line
+
+export interface PupilSearchFilterState extends FilterState {
+	tab: string;
+}
