+++ src/assignment/hooks/assignment-pupil-tabs.tsx
@@ -1,25 +1,23 @@
-import React, { useCallback, useMemo, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-
 import { TabPropsSchema } from '@viaa/avo2-components/dist/esm/components/Tabs/Tab/Tab';
 import { IconNameSchema } from '@viaa/avo2-components/src/components/Icon/Icon.types';
 import { Avo } from '@viaa/avo2-types';
+import { useCallback, useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import { ASSIGNMENT_RESPONSE_CREATE_UPDATE_TABS } from '../assignment.const';
 
 export function useAssignmentPupilTabs(
-	assignment?: Avo.Assignment.Assignment_v2
+	assignment: Avo.Assignment.Assignment_v2 | undefined,
+	tab: ASSIGNMENT_RESPONSE_CREATE_UPDATE_TABS,
+	setTab: (newTab: string) => void
 ): [
 	TabPropsSchema[],
 	ASSIGNMENT_RESPONSE_CREATE_UPDATE_TABS | undefined,
-	React.Dispatch<React.SetStateAction<ASSIGNMENT_RESPONSE_CREATE_UPDATE_TABS>>,
+	(newTab: ASSIGNMENT_RESPONSE_CREATE_UPDATE_TABS) => void,
 	(id: string | number) => void
 ] {
 	const [t] = useTranslation();
 
-	const [tab, setTab] = useState<ASSIGNMENT_RESPONSE_CREATE_UPDATE_TABS>(
-		ASSIGNMENT_RESPONSE_CREATE_UPDATE_TABS.ASSIGNMENT
-	);
 	const tabs: TabPropsSchema[] = useMemo(
 		() =>
 			[
@@ -61,5 +59,5 @@ export function useAssignmentPupilTabs(
 		[setTab]
 	);
 
-	return [tabs, tab, setTab, onTabClick];
+	return [tabs, tab as ASSIGNMENT_RESPONSE_CREATE_UPDATE_TABS | undefined, setTab, onTabClick];
 }
