--- src/assignment/hooks/get-assignment-by-id.tsx
@@ -1,15 +1,0 @@
-import { useQuery, QueryResult } from '@apollo/react-hooks';
-
-import { Avo } from '@viaa/avo2-types';
-
-import { GET_ASSIGNMENT_BY_UUID } from '../assignment.gql';
-
-export function useGetAssignmentById(
-	assignmentId: string
-): QueryResult<{ app_assignments_v2: Avo.Assignment.Assignment_v2[] }, { id: string }> {
-	return useQuery(GET_ASSIGNMENT_BY_UUID, {
-		variables: {
-			id: assignmentId,
-		},
-	});
-}
