+++ src/search/components/index.ts
@@ -1,3 +1,4 @@
 export { default as SearchFilterControls } from './SearchFilterControls';
 export { default as SearchResultItem } from './SearchResultItem';
 export { default as SearchResults } from './SearchResults';
+export { default as SearchFiltersAndResults } from './SearchFiltersAndResults';
