+++ src/search/search.const.ts
@@ -27,3 +27,15 @@ export const DEFAULT_SORT_ORDER: SortOrder = {
 	orderProperty: 'relevance',
 	orderDirection: 'desc',
 };
+
+export const GET_SEARCH_ORDER_OPTIONS = (t: (key: string) => string) => [
+	{ label: t('search/views/search___meest-relevant'), value: 'relevance_desc' },
+	{ label: t('search/views/search___meest-bekeken'), value: 'views_desc' },
+	{ label: t('search/views/search___uitzenddatum-aflopend'), value: 'broadcastDate_desc' },
+	{ label: t('search/views/search___uitzenddatum-oplopend'), value: 'broadcastDate_asc' },
+	{
+		label: t('search/views/search___laatst-toegevoegd'),
+		value: 'createdAt_desc',
+	},
+	{ label: t('search/views/search___laatst-gewijzigd'), value: 'updatedAt_desc' },
+];
